Separate server and frontend utils and extend confirm dialog functionality
