+++ components/ListView/index.tsx
@@ -3,7 +3,7 @@ import List from '@material-ui/core/List';
 import { Dispatch, FC, SetStateAction, useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import { buttonTheme } from '../../lib/themes';
-import { useConfirmDialog } from '../../lib/util';
+import { useConfirmDialog } from '../../lib/utils-frontend';
 import ListElement, { ListEl } from '../ListElement';
 import Popup from '../Popup/index';
